Resolve MSVC warnings for cleaner builds
Address MSVC warnings (C4005, C4018, C4189, C4267)

Changes:
- Remove conflicting CHRONO_VERSION macro definition in logging.h; add fallback in logger_backend.cpp to avoid conflict with Chrono's ChVersion.h (C4005)
- Use Eigen::Index for Eigen container loops instead of int/size_t (C4018)
- Use size_t for STL container loops in demos/tests (C4018)
- Add explicit casts for size_t-to-int conversions where APIs require int (C4267)
- Remove unused variables: rest_length, col_dof, total_dofs, bodies, dampers,  samples, width_array (C4189)
- Fix variable shadowing: rename t_tau to t_tau_init in ExcitationConvolution
- Mark Irrlicht includes as SYSTEM in CMakeLists.txt to suppress warnings

Build tested: Release configuration, all (working) regression tests pass.

+++ demos/f3of/demo_F3OF_DT1.cpp
@@ -202,7 +202,7 @@ int main(int argc, char* argv[]) {
                    << std::right << std::setw(16) << "Base Pitch (radians)" << std::right << std::setw(16)
                    << "Flap Fore Pitch (radians)" << std::right << std::setw(16) << "Flap Aft Pitch (radians)"
                    << std::endl;
-        for (int i = 0; i < time_vector.size(); ++i)
+        for (size_t i = 0; i < time_vector.size(); ++i)
             outputFile << std::left << std::setw(10) << std::setprecision(2) << std::fixed << time_vector[i]
                        << std::right << std::setw(16) << std::setprecision(4) << std::fixed << base_surge[i]
                        << std::right << std::setw(16) << std::setprecision(4) << std::fixed << base_pitch[i]
